Got the main moment's animation for the 60 LED strip!!
