+++ Core/Src/neopixel.c
@@ -1,0 +1,2180 @@
+/**
+  ******************************************************************************
+  * @file           : neopixel.c
+  * @brief          : Function and table definitions for driving the NeoPixels.
+  ******************************************************************************
+  * @attention
+  *
+  * Copyright (c) 2023 STMicroelectronics.
+  * All rights reserved.
+  *
+  * This software is licensed under terms that can be found in the LICENSE file
+  * in the root directory of this software component.
+  * If no LICENSE file comes with this software, it is provided AS-IS.
+  *
+  ******************************************************************************
+  */
+ 
+/* Includes ------------------------------------------------------------------*/
+#include "neopixel.h"
+
+
+/* Private typedef -----------------------------------------------------------*/
+
+
+
+/* Private define ------------------------------------------------------------*/
+
+
+
+/* Private macro -------------------------------------------------------------*/
+
+
+
+/* Public variables ---------------------------------------------------------*/
+const struct LED_Strip_Frame_S LED_ANIMATION_TABLE[NUM_OF_FRAMES] =
+{
+    {
+        .LEDs =
+        {   // 0
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 1
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 2
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 3
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 4
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 5
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 6
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 7
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 8
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 9
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 10
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 11
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 12
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 13
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 14
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 15
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 16
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 17
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 18
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 19
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 20
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 21
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 22
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 23
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 24
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     }
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 25
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 26
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 27
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 28
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+        },
+    },
+
+    {
+        .LEDs =
+        {   // 29
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=0,   .rgb.red=0,   .rgb.blue=0     },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+            { .rgb.green=255, .rgb.red=255, .rgb.blue=255   },
+        },
+    },
+};
+
+
+/* Private variables ---------------------------------------------------------*/
+
+
+
+/* Private function prototypes -----------------------------------------------*/
+
+
+
+/* Public function definitions -----------------------------------------------*/
+
+
+
+/* Private function definitions ----------------------------------------------*/
+
